Unify format helpers: dateTime and github names
