--- src/js/services/github.js
@@ -1,5 +1,0 @@
-export const getOrgName = fullName => fullName.split('/')[1];
-
-export const getRepoName = fullName => fullName.split('/')[2];
-
-export const getUserName = fullName => fullName.split('/')[1];
